gtksink: plug widget leaks
Calls to gst_gtk_base_sink_get_widget must unref the returned widget
when they are done with it. An internal reference is added when a new
widget is created, but a new reference is also added for each call to
this function, and the caller is responsible for freeing it.